pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

# The project name "lute" is already taken on PyPi,
# so project.name = lute3 for the publish, but the
# code is in the "lute" folder.
[tool.flit.module]
name = "lute"

[project]
name = "lute3"

dynamic = ['version']

description = "Learning Using Texts"
requires-python = ">=3.8"
authors = [
  {name = "Jeff Zohrab", email = "jzohrab@gmail.com"}
]
readme = "README_PyPi.md"

dependencies = [
  "Flask-SQLAlchemy>=3.1.1,<4",
  "Flask-WTF>=1.2.1,<2",
  "natto-py>=1.0.1,<2",
  "jaconv>=0.3.4,<1",
  "platformdirs>=3.10.0,<4",
  "requests>=2.31.0,<3",
  "beautifulsoup4>=4.12.2,<5",
  "PyYAML>=6.0.1,<7",
  "toml>=0.10.2,<1",
  "waitress>=2.1.2,<3",
  "openepub>=0.0.8,<1",
  "pyparsing>=3.1.4",
  "pypdf>=3.17.4",
  "subtitle-parser>=1.3.0",
  "ahocorapy>=1.6.2",
  "anthropic>=0.18.0,<1",
  "python-dotenv>=1.0.0,<2"
]

[project.scripts]
lute = "lute.main:start"

[package_data]
lute.config = ["config/config.yml"]

[project.optional-dependencies]
dev = [
  "coverage>=7.3.1,<8",
  "invoke>=2.2.0,<3",
  "pip>=23.0.1",
  "pipdeptree>=2.13.0,<3",
  "pylint>=2.17.5,<4",
  "pytest-bdd>=7.0.0,<8",
  "pytest-splinter>=3.3.2,<4",
  "playwright>=1.39.0,<2",
  "pre-commit>=3.5.0,<4",
  "black>=23.10.1,<24",
]

[project.urls]
Home = "https://github.com/luteorg/lute-v3"
